% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elicitPrior}
\alias{elicitPrior}
\title{Prior elicitation for Global Local Spatial Selection with varying coefficients}
\usage{
elicitPrior(
  parGrid0,
  t01,
  t02,
  alpha0,
  parGrid1,
  t11,
  t12,
  alpha1,
  parGridOm,
  iloc,
  Ae = 2,
  nu = 0.5,
  N = 1000
)
}
\arguments{
\item{parGrid0}{grid of parameters \eqn{(k, \phi_{\beta,0})} for the global selection when \eqn{\delta}=0}

\item{t01}{threshold for W_1 when \eqn{\delta}=0}

\item{t02}{threshold for W_2 when \eqn{\delta}=0}

\item{alpha0}{probability of event when \eqn{\delta}=0}

\item{parGrid1}{grid of parameters \eqn{(b_\eta, \phi_{\beta,1})} for the global selection when \eqn{\delta}=1}

\item{t11}{threshold for W_1 when \eqn{\delta}=1}

\item{t12}{threshold for W_2 when \eqn{\delta}=1}

\item{alpha1}{probability of event when \eqn{\delta}=1}

\item{parGridOm}{grid of parameters \eqn{(\xi^2, \phi_{\omega})} for the local selection}

\item{iloc}{(\code{n x 2}) matrix with the locations}

\item{Ae}{prior parameter for \eqn{\eta^2\sim IG(a_\eta,b_\eta)}; default 2}

\item{nu}{smoothness of spatial correlation kernel; default 0.5}

\item{N}{number of draws for the simulation-based elicitation; default 1000}
}
\value{
Elicited prior parameters and empirical probabilities (see paper for more details)

\code{prBeta}: a vector with \eqn{(b_\eta, k, \phi_{\beta,1}, \phi_{\beta,0}, \hat{\alpha}_0,\hat{\alpha}_1)}

\code{prGam}: a vector with \eqn{(\xi^2, k, \phi_{\omega}, \hat{p}_s, \hat{p}_{\rho|s})}
}
\description{
Prior elicitation for Global Local Spatial Selection with varying coefficients
}
