\name{GLoSS-package}
\alias{GLoSS-package}
\alias{GLoSS}
\docType{package}
\title{
\packageTitle{GLoSS}
}
\description{
\packageDescription{GLoSS}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{GLoSS}

Maintainer: \packageMaintainer{GLoSS}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = GLoSS)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{GLoSS}}
%\section{Documentation Index}{\packageIndices{GLoSS}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
