% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_gloss.R
\name{predictBetaGamma}
\alias{predictBetaGamma}
\title{Spatial prediction for Global Local Spatial Selection with varying coefficients}
\usage{
predictBetaGamma(j, mod, prParam, iloc_test, iloc_train, setting = "GLoSS")
}
\arguments{
\item{j}{index of the coefficient to predict}

\item{mod}{output of GLoSS function}

\item{prParam}{list of elicited hyperparameters (see paper for details)}

\item{iloc_test}{(\code{n_test x 2}) matrix with the new locations}

\item{iloc_train}{(\code{n_train x 2}) matrix with the locations used for estimation}

\item{setting}{one of the special cases GLoSS, V-LoSS, C-LoSS, GP; default is GLoSS}
}
\value{
Samples from posterior predictive distributions of the j-th coefficient and local indicator
}
\description{
Spatial prediction for Global Local Spatial Selection with varying coefficients
}
